Add a Windows build target to Travis checks and fix all problems it identified.

+++ src/hashtable.c
@@ -36,19 +36,19 @@
 hashtable_t *_hashtable_new(int size)
 {
     hashtable_t *t;
-    int size2, bits2;
+    unsigned size2, bits2;
 
     /* Adjust requested size to account for max load factor. */
     size = 1 + size * HASHTABLE_LOADFACTOR_DEN / HASHTABLE_LOADFACTOR_NUM;
     /* Use next power of 2 larger than the requested size and get mask bits. */
-    for (size2 = 2, bits2 = 1; size2 < size; size2 <<= 1, bits2++) ;
+    for (size2 = 2, bits2 = 1; (int)size2 < size; size2 <<= 1, bits2++) ;
     if (!(t = calloc(1, sizeof(hashtable_t)+ size2 * sizeof(unsigned))))
         return NULL;
     if (!(t->etable = calloc(size2, sizeof(void *)))) {
         _hashtable_free(t);
         return NULL;
     }
-    t->size = size2;
+    t->size = (int)size2;
     t->count = 0;
     t->tmask = size2 - 1;
 #ifndef HASHTABLE_NBLOOM
@@ -56,7 +56,7 @@ hashtable_t *_hashtable_new(int size)
         _hashtable_free(t);
         return NULL;
     }
-    t->bshift = sizeof(unsigned) * 8 - bits2;
+    t->bshift = (unsigned)sizeof(unsigned) * 8 - bits2;
     assert(t->tmask == (unsigned)-1 >> t->bshift);
 #endif
 #ifndef HASHTABLE_NSTATS
